+++ manager/app/src/main/java/me/weishu/kernelsu/ui/component/material/SearchBar.kt
@@ -56,8 +56,6 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.launch
-import me.weishu.kernelsu.ui.component.LocalSnackbarHost
-import me.weishu.kernelsu.ui.component.M3SnackBarHost
 
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalMaterial3ExpressiveApi::class)
 @Composable
@@ -225,7 +223,7 @@ fun SearchAppBar(
                 } else {
                     defaultContent(bottomPadding, collapseAndClear)
                 }
-                M3SnackBarHost(
+                SnackBarHost(
                     hostState = snackBarHostState,
                     modifier = Modifier
                         .align(Alignment.BottomCenter)
